Update tag_game_presets.json by Nimble102
Another bunch of presets: This time primarily located in Queen Gardens. Added Jordy's suggestion of Howling Cliffs/Gorb preset and reverted the minimum player count of his white palace preset back to 8 from 10 as per his request. Though, I still feel like the preset is a bit big for 8 players, so it is up to you if you want to keep this change.

Everything should be good, even though the Github is a bit different from how the indents, etc are set out in notepad. Tested the presets with the notepad version but not the Github version though, and they were working there.